Show postal confirmation message at booking stage
This was previously only being displayed at the appointment stage.
