+++ app/views/appointments/edit.html.erb
@@ -69,7 +69,7 @@
       </div>
 
       <% if @appointment_form.postal_confirmation? %>
-      <hr>
+        <hr>
       <h2 class="h3">Postal address</h2>
       <p>
         <%= simple_format(@appointment_form.postal_address_lines) %>
